Fixed mini game add coin (BY BEN) , fixed setting state ( buttons) , changed time , fixed choice

+++ trunk/Project/Resource.h
@@ -4,18 +4,18 @@ class CResource
 
 private:
 	static int NumOfFood;
-	static float NumOfMoney;
+	static int NumOfMoney;
 	static int NumOfManPower;
 	static int NumOfCitizen;
 public:
 
 	int GetFood();
-	float GetMoney();
+	int GetMoney();
 	int GetManPower();
 	int GetCitizen();
 
 	void SetFood(int NewFood);
-	void SetMoney(float NewMoney);
+	void SetMoney(int NewMoney);
 	void SetManpower(int NewManPower);
 	void SetCitizen(int NewCitizen);
